pyproject.toml
[project]
name = "cang-cli"
version = "0.1.0"
description = "仓咚咚 - 个人金融命令行工具 (AI Native)"
readme = "readme.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "gudong"}
]
dependencies = [
    "typer>=0.12.0",
    "dong-core>=0.1.0",
]

[project.scripts]
cang = "cang.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cang"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "-v",
    "-l",
]
console_output_style = "count"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "unit: 单元测试（测试单个函数/类）",
    "integration: 集成测试（测试 CLI 命令）",
    "slow: 慢速测试（涉及 I/O 或大量数据）",
    "db: 数据库相关测试",
    "cli: CLI 命令相关测试",
]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src/cang"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
